Implement initial improvements
* Ensures that TransifexNative SDK can run on both Swift and Objective-C
projects on iOS 10.3+.
* Improves the documentation of all classes and methods.
* Improves syntax in methods by using `guard` and optional statements
where applicable.
* Uses a backwards compatible `md5()` method so that it can be used in
older iOS versions.
* Improves the library initialization logic.
* Refactors classes and interfaces.
* Implements asynchronous logic in `CDSHandler` class with logic that
deals with 202, 200 and error responses.
* Refactors `SourceString` into a `struct` that can be encoded and
decoded easily (`Codable`).
* Creates first unit tests that test different classes and methods.
* Updates README to include library initialization for Objective-C
projects.
* Refactors `getHeaders` method to use a guard statement.
* Makes `tags` property of `SourceStringMeta` optional.
* Updates `debugDescription` of both `SourceStringMeta` and `SourceString`
classes.
* Adds `CDSHandler` tests by refactoring `URLSessionMock` to accept
an array of mock responses `MockResponse` and creates a test for the
case when the results are not ready on CDS and two `202` status codes
are returned before the `200` one.
* Fixes trailing slash issue in `cdsHost` property by transforming the
`cdsHost` argument to a `URL` and appending the components using the
`URL` provided methods.
* Introduces `CurrentLocaleProvider` protocol so that the current locale
of the application can be fed via a separate class that conforms to said
protocol.
* Offers `UserDefaultsLocaleProvider` as a first class that implements
the above protocol and returns the current application locale as found
in `UserDefaults` dictionary of the app.
* Introduces a new test that tests the `CurrentLocaleProvider`
functionality and how it interfaces with `LocaleState`.
* Renames `Rendering.swift` to `Policies.swift` to better reflect the
contents of that file.
* Applies `final` keyword to whichever classes need to be considered
final.