Implement swizzling logic
Implements the swizzling logic in order to intercept all of the
`NSLocalizedString()` and `localizedStringWithFormat()` calls both in
Swift, Objective-C and Storyboard files.

The logic is wrapped inside a `Swizzler` class that's responsible for
activating the Swizzling in all of the aforementioned cases.

For the `NSLocalizedString()` and storyboard / xib files case, the
`SwizzledBundle` class is used that swizzles all application bundles so
that the `localizedString(forKey:value:table:)` is overriden and
intercepted by the Swizzler class.

For the `localizedStringWithFormat` case in Swift, the interception
occurs via extensions for both the `String` and the `NSString` classes
that call back the respective `Swizzler` method. This extensions file
(`TXNativeExtensions.swift`) needs to be included in the project that
uses the TransifexNative package and added to all of the necessary
targets so that the extensions can intercept those method calls.

For the `localizedStringWithFormat:` case in Objective-C, a separate
Objective-C class had to be created (as swizzling a variadic method in
Swift is not supported) and the swizzled method had to properly parse
the variadic arguments. This happens via a regular expression that
extracts all format specifiers and then manually detects the type (as
specified by Apple and the C spec) and passes that type to the va_arg()
function that extracts the argument. Finally, all of those arguments are
converted to Objective-C objects (`NSNumber`, `NSString` etc) and are
passed back to the `Swizzler` class via an array which can be properly
enumerated.

Unit tests have been implemented for the `TXNativeObjSwizzler` class.

Adds platform strategy as an enum in TransifexNative initialization
logic, that defaults to the platform strategy.

Adds simple logic that extracts the pluralized ICU format, parses the
'one' and 'other' rules and applies the proper rule depending on the
value of the extracted argument.

Moves the serial dispatch queue of the `MemoryCache` class from the
`NativeCore` class to the `MemoryCache` and uses it to both update and
read values from using the `sync` method, so that it's ensured that all
read / write operations happen in a serialized fashion.

Also updates `Cache` protocol `update` method so that a whole dictionary
of locale codes can be provided in order for the internal data structure
to be updated in one dispatch call.

An empty `init()` constructor has been implemented in the
`URLSessionMock` class to silence two deprecation warnings.