Add format rule extraction logic
Adds sample `Localizable.stringsdict` file that is used to extract which
rule to choose when performing the pluralization.

Implements parsing of ICU Message Format:

* Extracts pluralized phrases from ICU strings.
* Supports strings like the following:
{???, plural, one {One table} two {A couple of tables} other {%d tables}}

Note: variable placeholders that look like "{var}" are not supported,
as they need a more complex grammar than simple regular
expressions.

Uses provided locale in rendering

Refactors `format()` method of PlatformFormat so that's more readable
and it properly falls back to the default string interpolation if no
format can be picked by the rule extraction logic.

The method now falls back to the "other" and not to the "one" rule when
necessary, as it's the only rule that all locales always have.