Add push functionality to CDS Handler
Adds functionality to push base translations to CDS by introducing the
`pushTranslations()` method in `TxNative`, `NativeCore` and `CDSHandler`
classes.

Alongside this method, the `TxSourceString` class is also introduced
which offers a way for both Swift and Objective-C projects to construct
the data structure that will hold all the necessary information for the
strings to be pushed. The `TxSourceString` also offers an internal
method that transforms the object to the `Codable` `SourceString` struct
that can be used when constructing the data to be pushed.

This commit also moves the `occurrences` property from `SourceString`
struct to `SourceStringMeta`, where it should have been originally
placed.

The `pushTranslations()` method of `CDSHandler` serializes the passed
array of `TxSourceString` objects using an internal `PushData` struct
and the resulting `Data` object is passed to the `URLRequest` that is
used to push the strings to CDS. An extra optional `purge` parameter can
also be used to replace the entire resource content, if needed.

When the push logic completes, a completion handler informs the caller
asynchronously on whether the request was successful or not.

This commit also adds a unit test that tests whether the provided
`TxSourceString` objects are properly transformed into the expected
JSON representation before being pushed to the CDS server.

Finally this commit removes a couple of unused `;` characters and
`break` statements.