Refactor locale provider logic
* Renames `CurentLocaleProvider` to `TXCurrentLocaleProvider` so that
there are no potential conflicts in Objective-C projects.
* Renames `UserDefaultsLocaleProvider` to `TXPreferredLocaleProvider`
and fetches the app's current locale from the `NSLocale` API using the
`autoupdatingCurrent` property instead of looking at `NSUserDefaults`.
The logic still falls back to the "en" language code, if the language
code cannot be extracted from the `autoupdatingCurrent` property and
also listens for the `NSLocale.currentLocaleDidChangeNotification` in
case the locale changes during runtime. Typically this notification is
never called, as the app always restarts when the locale is changed by
the user from the Settings app (either the per-app locale or the global
device locale).
* Updates documentation of `LocaleState` constructor regarding default
values.
* Introduces `translatedLocales` property that returns the list of
locales supported by the application, without the source locale. This
property is used by the `CDSHandler` on it's initialization, so that
only those locales are fetched.
* Removes invalid comment from the documentation of the `NativeCore`
constructor.
* Updates README to include the source locale in the list of
`appLocales` in the `LocaleState` examples and also updates the line
breaks to 80 column length.
* Ensures that the source locale is always added in the `appLocales`
list of the `LocaleState` even if it hasn't been added by the caller.