Improve pull logic, refactor cache
The `fetchTranslations` method now offers a completion handler so that
the caller is notified when the downloading of the translations is
complete.

The completion block of that method now also reports any errors that
may have been encountered during the download operation so that the
caller can decide how the resulting translations should be treated.

`LocaleStrings` and all of the depended type aliases have now a `TX`
prefix to avoid any unwanted naming collisions.

Caching logic has been improved: If an object that conforms to the `TXCache`
protocol (which has also been renamed) isn't provided during the
`TxNative` initialization, then the internal `TXDefaultCache` is used,
which uses the decorator pattern to create two `TXCacheProvider` objects
and feed them to the `TXProviderBasedCache` which acts as the internal
cache of the `TXDefaultCache`. The `TXProviderBasedCache` uses a
`TXMemoryCache` to store all the extracted translations in memory and
for reporting them back via the `get()` method of the `TXCache`
protocol.

The `TXCacheProvider` objects are two `TXDiskCacheProvider` instances
that try to look up translations both in the main app bundle and in the
application directory to find any already generated or downloaded
translation files accordingly.

On top of that a `TXNoOpCache` has been also implemented for apps that
don't want to initialize the `TxNative` library with a cache mechanism
(e.g. the cli tool).

When the new translations are fetched from CDS, the downloaded file is
either generated or replaced by those new translations.

Also, for applications that make use of several different targets (e.g.
extensions), the `TXDefaultCache` can be initialized with the optional
`groupIdentifier` argument, so that the downloaded translations are
stored in the application group in order for all of the targets (e.g.
extensions) to be able to have access to them.