Implement decorator classes for cache
Implements decorator classes for the cache module based on the
discussion on the Github pull request.

Updates README with details regarding the standard cache, also detailing
alternative uses. Adds a section in README on using the SDK in
application extensions.

Updates unit tests to reflect on the latest cache changes.