Public release
* Adds logging support and exposes logging protocol, enums and helper
classes for developer use.

Adds logging support in SDK by exposing `TXLogger`, log levels and
`TXStandardLogHandler` so that developers can easily specify a desired
minimum log level. By default, the minimum log level is set to `.warning`
which logs warning and error log statements.

* Updates all public symbols, improves documentation and adds `TX`
prefix

Adds `TX` prefix to all publicly exposed classes, protocols, enums,
typealiases and methods so that they don't conflict with any Apple
related APIs on Objective-C projects.

Renames `TxNative` to `TXNative`.

Updates documentation and visibility of certain methods.

* Updates README with instructions and limitations.
* Adds error policy unit test

Makes the `ICUMessageFormat` rendering strategy to `throw` an error when
its `format()` method is called, so that the error policy will be
triggered by the `Core` class.

Adds related unit tests.

* Fixes `fetchTranslations` issue with no locale codes

The `completionHandler` of the `fetchTranslations` method of the
`CDSHandler` was never called if the method was called with no locale
codes.

This commit fixes the issue by checking the `fetchLocaleCodes` list size
and if it's empty, it logs the error and calls the `completionHandler`
with the proper error code.

* Removes default `ios` tag from `SourceString`

Removes the default "ios" tag from the `SourceString` logic if no tags
are specified.

* Implements logic for bundle translations

Implements `BypassLocalizer` internal class that's responsible for
looking up on the application bundle for the source translations, if
available, so that if the application runs on the source language or if
the missing policy is about to be used, the source translations are
used.

This is done by loading the bundle for the source language and skipping
the swizzling logic so that the native localized string calls can fetch
the base localization.

If the base localization resources cannot be retrieved, then the logic
falls back to the value provided by the developer (if available) and if
not, it falls back to the source string.

* Filter empty strings in override all policy

Filters out empty translation strings in all policies by introducing a
separate method that's responsible for filtering the `TXTranslations`
structure.

Adds unit test that checks if that filtering occurs in the override all
policy.

* Exposes CDS cache invalidation as a public method
(`forceCacheInvalidation:`).
* Improves logging of cache invalidation.
* Improves README by offering more explanation regarding the
`fetchTranslations:` method as well as renaming the "Transifex Native"
to "Transifex" where needed.

* Renames the Swift Package from TransifexNative to Transifex

* Refactor of translate method regarding plurals

The `translate(sourceString:params:)` method which is responsible for
rendering plurals, was also going through the
`translate(sourceString:localeCode:params:context:)` -> `render`
call stack in the plural string case without needing to.

This change makes it so that the `translate(sourceString:params:)` calls
the `render` method directly as we don't need to resolve the key for the
`sourceString` in the plural case.

* Log error response when pushing source strings

Adds logging for CDS response text when there is an error while pushing
source strings.

* Makes the `secret` parameter of `TXNative` constructor optional in
Swift.

* Always update cache on fetchTranslations completion

Cache is always updated when the `fetchTranslations` of the `CDSHandler`
class completes, regardless of whether there are errors or not.

This is done because the user might have requests translations for
locales that are not yet supported and those locales would trigger 404
errors from CDS, blocking the normal translations from the locales that
are available, to update the internal cache, and by extension, to be
saved to the application sandbox directory.

* Adds `CHANGELOG.md`.
* Changes workflow name from `Swift` to `CI`.
* Adds badges in README.
* Adds minimum requirements table in README.
* Fixes typos and code comments.

* Introduce a simpler SDK initialization method

Introduces the `initialize(locales:token:)` method for a faster SDK
initialization, using the default values for the rest of the arguments.

* Handles 403 - Forbidden error while pushing strings by immediately
returning `false` instead of trying to decode the response data as a
JSON.
* Converts a logger `error()` to `warning()` when loading a file from the
file system cache as it's expected to be missing the first time the app
launches.