Refactor cache policies
Refactors cache policies based on the latest discussion between
@dbendilas and @petrakeas.

The empty translations (`""`) are now added in the cache (depending on
the policy) and they are not filtered out.

A helpful cheat-sheet table has been added in the documentation of the
`TXCacheOverridePolicy` and the policy documentation has been also
updated.

The cache now defaults to the `.overrideUsingTranslatedOnly` instead of
the `.overrideAll` policy, if no policy is specified by the developer.

If a cache entry is found to contain an empty string during rendering,
the logic falls back to the missing policy.

The documentation of the `TXDiskCacheProvider` has been updated to
declare that it initializes the disk cache provider synchronously.

* Improves policy unit tests by using the table provided in the
documentation to initialize a memory cache and a set of new
translations.

* Adds a unit test that tests whether the read only cache
(`TXReadOnlyCacheDecorator`) is immutable.

* Refactors `TXStandardCache` to use a static method that constructs and
returns the `TXCache` object instead of using the decorator pattern.

Updates the default standard cache usage in the `NativeCore` class and
the README instructions.

* Renames override policy enum to update policy and updates the values to
match the ones used by the Java SDK.

* Replaces occurrences of the 'override' term (relating to the override
policy) with the 'update' term, to match the rename of the override
policy enum.

* Updates CHANGELOG, bumps version