Allow the pull command to fetch source string content
Given that CDS now allows the pulling of source content, changes had to
be made to enable that functionality on the SDK.

This commit changes the `CDSHandler` initialization by allowing it to
accept all application locales that have been declared by the developer
during the SDK initialization, which include the source locale.

Updates translation logic to correctly support source and target locales
and cover cases where the provided source string is not found on cache
correctly.

The missing policy in the `translate()` method is now applied after the
`render()` so that any transformations by the missing policy will not
affect the string interpolation.

A new test has been added to test whether the source locale is fetched
from the cache for a simple and a plural approach.

* Introduces a public static `dispose()` method in the `TXNative` class,
so that the `TXNative` singleton instance can be safely deconstructed in
order for another `TXNative` instance to be initialized and used.
* Calls the aforementioned `dispose()` method when each test ends
(`tearDown`) so that the next unit test can properly initialize and use
a new `TXNative` instance.