Fetch content with tags only
Adds an optional argument to `TXNative.fetchTranslations` method
(`tags`) that allows users to define a set of tags to use when fetching
translations from CDS. If defined, CDS will only return content for
strings that have all of the given tags.

Adds unit test that expects the final URL sent to server to have a
specific format.

Ref:
https://trello.com/c/eWgaL5BB/49-fetch-content-with-tags-only-on-runtime