Introduce activation method for Swift modules
When an application is built with a modular approach using Swift
Packages, any localization logic existing in those modules could not be
intercepted by the SDK, as the module bundle is not included in the
`.allBundles` property when activating the `SwizzledBundle` class.

In order for that to happen, a new method (`activate(bundle:)`) has been
introduced in the public API of the SDK that allows developers to pass
the `Bundle.module` of each Swift package that needs to have its
localization logic work with the SDK.

Developers must ensure to call this method from each of the Swift
Packages after the TXNative SDK has been initialized.