Implement proper tag and status filter support
Adds support for filtering the fetched translations either by a list of
tags and/or by status.

The filtering can be either specified globally on the TXNative
initialization by passing the `filterTags` and/or the `filterStatus`
properties, or it can be specified every time the `fetchTranslations()`
method is called. In the latter case, the passed tags or status will
have a priority over any already specified tags or status specified
during initialization.

The `session` property in the `TXNative` initialization is now used
correctly.

The unit tests have been updated to test both the `TXNative` and the
`CDSHandler` behavior on translation filtering using tags and status.

Due to the number of different arguments on the `CDSHandler`
initialization, a separate struct was created to hold this information
called `CDSConfiguration`. This struct is initialized and passed
internally during the initialization of the `CDSHandler` by the SDK.