Improve CDSHandler push logic
The `pushTranslations()` method now returns a secondary array of `Error`
objects in the completion handler that represents the generated warnings
during the processing of the source strings that are non-fatal but can
be of interest to the developer.

This change introduces a new enum (`TXCDSWarning`) that hosts those
warnings, like duplicate source string and empty key detection.

The `pushTranslations()` logic has been improved by implementing the
following changes:

* `serializeTranslations()` is now a private static method as it does
not need to look into the state of the `CDSHandler` instance.
* That method now returns a tuple that includes a `Result<Data>` object
as the first member, revealing whether the serialization
was successful, thus offering the `Data` object, or a failure, thus
offering a `Error` object for further propagation and an array of
`TXCDSWarning` objects for any generated warnings during the processing
of the source strings.
* The `pushTranslations()` implementation has been split so that it's
easier to be read. The new method that performs the asynchronous network
operation is now separate (`pushData()`) and propagates the generated
warnings to the `pollJobStatus()` method.
* Any error logging has been removed in favor of the list of `Error`
objects in the completion handler. The logging is now a responsibility
of the consumer.