Address language tag discrepancy
iOS accesses the language tags using a hyphen (e.g. `en-GB`), whilst
Transifex uses an underscore for the delimiter (e.g. `en_GB`).

This discrepancy was causing an issue in the `BypassLocalizer` class
logic, that is used to access the underlying bundled localized asset of
the app as a final fall-back when the translated strings cannot be found.

This fix replaces the underscore character (`_`) with a hyphen (`-`) in
the provided `localeCode` of the `BypassLocalizer` constructor, to
ensure that the underlying asset is always accessible by the logic.