Ensure that callbacks won't capture self strongly
Given that the main SDK instance (`tx`) can be set to `nil` using the
`dispose()` method at any time, we need to ensure that the callback
methods of the networking logic (`fetchTranslations`, `pushTranslations`,
`foceCacheInvalidation`) will not capture the instance strongly, retaining
it more than necessary.

In order to address this, the internal `CDSPullRequest` class has been
removed in favor of a private method in the `CDSHandler` class
(`performFetch(...)`) that behaves similarly and ensures that the
network callbacks are capturing a weak reference to the `CDSHandler`
instance.