Ensure Designed for iPhone/iPad apps use the proper device name
In device variation rules, the logic uses the iPad device name when the
user interface idiom is Pad,but that does not cover the corner case
where an iOS application targets a Vision device as a 'Designed for iPad'
application.

In order to on par with the logic implemented by Apple, the 'Designed
for iPad' applications on Vision devices, now use the correct device
name. The rest of the cases were also reviewed ('Designed for iPhone' /
'Designed for iPad' iOS apps running either on Mac or Vision devices).

More specifically:

* iPad applications running inside a Mac ("Designed for iPad") return
the "ipad" device name.
* iPhone applications running inside a Mac ("Designed for iPhone")
return the "iphone" device name.
* iPad applications running inside a Vision Pro ("Designed for iPad")
return the "applevision" device name.
* iPhone applications running inside a Vision Pro ("Designed for iPhone")
return the "iphone" device name.